implement handlers design and erase error handling
